[contract] Move Escrow Tests and Add New Test
